perf(Processing/Computations): Move computations utility functions under `utils`
