+++ processing/processing/utils/load_features_for.py
@@ -11,24 +11,24 @@ def load_features_for(cfg, band, r, s):
     range_times = []
     range_features = []
     for fname, info, audio, npz in iterate_audio_files(cfg, band,
-                                                        ['@feature_base',
-                                                         '.npz']):
+                                                       ['@feature_base',
+                                                        '.npz']):
         if info.site != s:
             continue
         if npz.exists():
             data = np.load(npz)['x']
-        else: # backward compatibility, try pklz
+        else:  # backward compatibility, try pklz
             with gzip.open(npz.with_suffix(".pklz"), "rb") as f:
                 data = pickle.loads(f.read())
             # TODO: proper error message (saying that npz also not found)
 
-        dur = dt.timedelta(seconds=0.92 * len(data))
+        dur = dt.timedelta(seconds=1 * len(data))
         if r[0] > info.start + dur:
             continue
         if r[1] < info.start:
             continue
         for i in range(len(data)):
-            start = info.start + dt.timedelta(seconds=0.92 * i)
+            start = info.start + dt.timedelta(seconds=1 * i)
             if start < r[0] or start > r[1]:
                 continue
             range_times.append(start)
